Use custom attrs types at all edges, for consistency
Summary:
Cleaned up the behavior around the use of custom types. Previously, we required callers to submit a dict to the Client, but the Client would return a `JobResponse`. This makes the Client accept `JobRequest`/`ActionRequest` objects everywhere.

This is preferable for two reasons: first, it is less confusing to use custom types everywhere than to return them but not accept them. Second, custom types make code more explicit and easy to understand. It gives an extra layer of certainty that a request is well-formed before it is sent.

Added a Client section and a section about switches to the README. The other sections are left as stubs for now.

Test Plan: Unit tests are updated to use both raw dicts and Request objects.

Reviewers: bartek, ashfall, #foundry

Reviewed By: bartek, ashfall, #foundry

Subscribers: ashfall, bartek, jenkinsbot, michaelmanganiello

Differential Revision: https://phabricator.evbhome.com/D32980