Change middleware to work in a callable (new-Django) style
Summary: This allows the middleware to use features like try/finally. It's based almost directly off of the Django 1.10 middleware style, but with .job and .action attributes rather than using the object as a callable directly, so we can still bundle middleware as a single class.

Test Plan: Manually tested

Reviewers: #unicode_reviewers, #foundry, bgreenberg

Reviewed By: #foundry, bgreenberg

Subscribers: seth, bgreenberg, jenkinsbot, michaelmanganiello

JIRA Issues: EB-55246

Differential Revision: https://phabricator.evbhome.com/D33831