Simplify settings definitions, and fix all the tests that were broken
Summary:
Simplifying settings:

Provide a schema for ASGI transport, and `ClientSettings` and `ServerSettings` subclasses that use it. Along with `ASGIClientSettings`, there is an `ASGIClientRouter` that uses it.

The other big settings change is that all SOA settings use `MsgpackSerializer` by default. This won't have any effect on existing code, because it all uses `msgpack` anyway, but since we have made `msgpack` our preferred serializer, we may as well default to using it.

Fixing tests:
In the course of making these changes, I noticed that a lot of tests had not been updated in a while and were failing. I've fixed the stub service code to work with the new `LocalTransport` interface, and fixed tests for schemas and service stubbing, as well as a handful of tests that were using the old request format.

I also fixed a few minor problems that I found along the way: putting `uincode_literals` in a few places where it was missing, converting some `unittest`-based tests to `pytest` and, channeling Nick, fixed some mixed quotes.

Test Plan: All unit tests now pass.

Reviewers: bartek, #foundry, #unicode_reviewers

Reviewed By: bartek, #foundry

Subscribers: bartek, jenkinsbot, michaelmanganiello

JIRA Issues: EB-56982, EB-57550

Differential Revision: https://phabricator.evbhome.com/D35615