Use master_for correctly to reduce number of Redis connections
For other Redis-consuming projects, we have traditionally used the `sentinel_sharded_redis` library, which has a `sentinel_refresh_interval` setting. This setting controls how often `sentinel_sharded_redis` reaches out to Sentinel to get the latest information about Redis masters.

When consuming the `redis-py` library in this project, we applied the same pattern, adding a `sentinel_refresh_interval` setting to the transport and using that to control how often `redis-py`'s `sentinel.master_for(...)` is called. This is not the way `sentinel.master_for(...)` was designed to be used.

`sentinel.master_for(...)` creates a whole new `StrictRedis` client with a whole new connection pool every time it is called. We are wasting connections that remain open and unused until garbage collection runs, and opening connections too often, by regularly calling `sentinel.master_for(...)`. This method was designed to be called in a startup or lazy-startup capacity, and the client it returns should be used indefinitely. The client it returns uses a special managed connection pool with special managed connections that keeps track of master information and automatically closes and re-opens connections if the master information changes.

In this change, we are removing the `sentinel_refresh_interval` setting (but it can still be included for a no-op for temporary backwards compatibility with existing settings). We will no longer call `sentinel.master_for(...)` regularly. Instead, we call it the first time we need a master client, and then we use that master client indefinitely.

- `master_for` behavior: https://github.com/andymccurdy/redis-py/blob/5109cb4f6b610e8d5949716a16435afbbf35075a/redis/sentinel.py#L246-L272
- Sentinel-managed connection pool behavior: https://github.com/andymccurdy/redis-py/blob/5109cb4f6b610e8d5949716a16435afbbf35075a/redis/sentinel.py#L99-L108
- Sentinel-managed connection behavior: https://github.com/andymccurdy/redis-py/blob/5109cb4f6b610e8d5949716a16435afbbf35075a/redis/sentinel.py#L33-L66
- An example of Flask using `redis-py` Sentinel the same way we're proposing in this change: https://github.com/exponea/flask-redis-sentinel/blob/9f0003345d3756c19ce913acb0e25b613f8368a6/flask_redis_sentinel.py#L62-L77