Remove the deprecated and unused ASGI Transport
The ASGI Transport, deprecated in November 2017 for performance issues and replaced with the now-default Redis Gateway Transport, is now removed as of 0.24.0 (January 2018). Additionally, several bits of documentation have been updated to reflect changes made to PySOA over the past six months, such as removal of the `ClientRouter` and the ASGI Transport.