Ensure that service name passed to client is always unicode
Everything else (settings, requests, responses, etc.) we enforce to be unicode. We don't enforce this for service names at the top level in the client, but we do at a lower level in the remote Redis transport. This results in hard-to-diagnose bugs that are often missed in testing (which uses a local transport).