Fix bug causing response mix-ups with transport cache
When the transport cache is enabled, separate clients on separate threads with identical transport settings could get their responses mixed up, such that "client 2" could get responses meant for "client 1" and "client 1" responses meant for "client 2." So, that's bad. This commit fixes that and also adds a test to confirm the problem and a test to confirm the fix.