Add support for auto-reloading code changes in dev environments
Using the `--use-file-watcher` option when starting a service with `example_service` or `python -m example_service.standalone` will now result in the service's being wrapped by a reloader that watches for file changes and restarts the service when file changes are detected. This feature is extremely useful for, and also only intended for, development environments. It is not recommended for production environments. It is fully compatible with both standard mode and forking mode (`--fork-processes` / `--fork` / `-f`).