Add support for clean-up operations before and after requests
If Django support is enabled in a server, we need to reset the query log after each request and close broken database connections before and after each request and when the server is idle. This replicates Django request processing behavior. In his change, it is implemented in such a way that implementing services can extend this behavior if necessary.