Add support for server introspection
This commit creates a new stock introspection action, which returns information about the service and its actions and their request and response schemas, and uses it by default if the server does not define an `introspect` action. As a related feature, it also ensures that every server responds to `status` if the server does not define a `status` action. Finally, it ensures the status action is adequately documented, so that it can be introspected well.