Fix unit test failures caused by Django connection cleanup
When unit tests are run, the Django test fixture transaction wrapping can break the PySOA server connection cleanup, manifesting as an strange atomic block error on the first database query within an action. This commit ensures we do not clean up connections if we're in the middle of a transaction, which solves that problem. It also ensures that connection cleanup failures caused by PyTest's `mark.django_db` test are ignored.