Ensure logging context works with local services by using a stack
The logging context added in 157b6e8 works great for remote services. It even works for calling a local service from something other than another service. However, when a local or remote service calls a local service (such as when unit testing a service that calls another service and stubbing that other service), it results in an attribute error in the outer call. To properly support local services, which could be called more than one level deep, the logging context should be a stack, not an absolute value, to ensure that we properly step into and out of each context for logging purposes, so that logging calls include the correct details for our current position in the nested PySOA call stack.