Make disable_existing_loggers default to False to allow module-level getLogger
When configuring Python logging with `dictConfig` (which the PySOA server does on startup), the `disable_existing_loggers` setting defaults to `True` for backwards compatibility with <2.6 versions of Python. This is true even in Python 3.6. The problem with this is that it is incompatible with code that instantiates a logger with `logging.getLogger` at the top of a module. When this code is first imported at startup, all loggers are initialized with a default config. When the new config is installed, all loggers not specifically mentioned in the new config are disabled (so they don't trickle to the root logger). This change sets `disable_existing_loggers` to `False` by default if not specified intentionally as `True`, which means that module-level `getLogger` calls triggered at import will not cause logging to break.

It should be noted that this change also mirrors Django behavior, which specifically defaults `disable_existing_loggers` to `False` unless the settings module has its value intentionally set to `True`.