[MAJOR] Switch to microsecond resolution for metrics timers
Due to the miniscule amount of time consumed by many operations in PySOA, millisecond resolution is not sufficient for metrics timers. This commit changes PySOA timers to microsecond resolution.

Note that this is a breaking change not for existing services/code, but for existing metrics graphs and reports. Previously, the value these graphs and reports displayed represented a number with milliseconds units. Now, they will be a number with microseconds units. As such, without the context of this change in mind, performance will appear to get worse by three orders of magnitude across the board on all existing graphs after a release deployment.