Add stock ability to include other services' status in status
This change adds the standard ability to include other services' `status` action response in the diagnostic information of the `status` action response of a service configured to call those other services. It also includes a small enhancement to `stub_action` necessary to fully test this change to the `BaseStatusAction`. Finally, it includes an enhancement to `BaseStatusAction` to pass the `request` object to `check_` methods.