Improve auto-docs using built-in method designed for it
The code was stripping to much indentation from auto-generated directive documentation, which was causing important contextual documentation to be lost. This uses a built-in method designed to remove leading and trailing blank lines and uniformly remove leading spaces from all remaining lines in Python docstrings so that contextual indentation is not lost.