Properly copy PyTest marks to fixture test cases
This change ensures that class-level PyTest markers get copied from `ServicePlanTestCase` classes to the fixture test case PyTest function objects so that fixtures, plugins, and the PyTest core properly handle the markers. This is necessary for, for example, `pytest.mark.django_db` to enable Django database access in test plans. As a positive side effect, we were able to get rid of some of the hacky skipping code in the plugin.