[MINOR] Remove the transport cache as it is no longer needed
The transport cache (controlled by the `transport_cache_time_in_seconds` setting), was put in place before we got our Redis connection handling under control, when creating new transports took multiples of milliseconds. Now that we properly handle Redis connections with a thread-safe, global connection cache, creating new transports take nanoseconds, the same as retrieving transports from the cache. Additionaly, the cache is not thread-safe and introduces problems in multi-threaded environments.

As such, we are removing the transport cache from PySOA. This is a backwards-compatible change, because we are simply deprecating the `transport_cache_time_in_seconds` setting, and leaving it in place for some time. There will be a future, breaking change that removes this setting altogether. Currently, its value is completely ignored.