[MAJOR] Add a response context dict to all responses
We already have a request context, which includes things like the correlation ID, session ID, locale, authentication/authorization information, etc. We now have a need of a response context for the service to send job-level response information back to the client. (This change also removes support for Python 3.3, as the build tools we use no longer support it.)

This is a backwards-incompatible change. All service clients must be upgraded before service servers are upgraded. When a client deserializes a response, it sends the response dict to `JobResponse(**response_dict)`. This will error with an unexpected keyword argument `context` if the client hasn't been upgraded.