[MINOR] Add option to suppress responses for send-and-forget
We've started using a send-and-forget concept, where we send a request but neither care about nor wait on a response. In these cases, the response the server sends needlessly sits in the Redis queue for 60 seconds until it expires. Using a `suppress_response` flag, you can now implement send-and-forget and instruct the server to not send a response, in order to conserve resources.