[MAJOR] Make the maximum Redis transport message size configurable
The maximum message size for the Redis Gateway transport was not set to a reasonable amount. This change makes it a configurable setting instead of a hard-coded value, and sets more reasonable defaults for clients and servers, separately. It also does some log refactoring to improve logging around message size. In particular:

- BREAKING: `pysoa.server.logging` has moved to `pysoa.common.logging` so that it can be used by the transport layer
- `MessageTooLarge` errors now include the message size
- The message size is logged when a messaged fails to send due to being oversized
- You can now configure a lower threshold at which to log, but still send, less-oversized messages