[MINOR] Add support for a heartbeat file
It's often advantageous to have the ability to programmitically check the health of a specific server process, which can't easily be achieved without using a client, and which can't be achieved with a client in a multi-server environment. This commit adds support for an optional (configured via settings) heartbeat file, which gets created when the server starts, updated after each request or during idle processes, and removed when the server shuts down.