[MINOR] Make the polymorphic client and server settings extensible
Before this change, if a consumer of PySOA (such as a service or another library) provides a new transport with a validated settings schema, the only way to use the transport and ensure the settings for it are validated is to also extend/override `PolymorphicServerSettings` and `PolymorphicClientSettings` and extend the `Client` to use the new transport settings class. This change makes the polymorphic client and server settings extensible by adding a `settings_schema` class attribute to a transport class.