[MINOR] Allow setting `side_effect` while defining the stub
Previously, it was only possible to set `side_effect` after the mock was returned by the `stub_action` context manager, which made it impossible to define preconfigured stubs (specially when setting a callable that rely on variable data). This change allows you to specify a `side_effect` when you create the stub.