[MINOR] A better Syslog logging handler
The Syslog logging handler built into Python has a crucial flaw: If the total, formatted log message exceeds the minimum interface MTU between the service and the Syslog server, that entire message is lost if sent over UDP (the default and best-performing option). It will either log the error or silently fail, but the message will be lost. Since PySOA logging includes request/response logging, which can be large, it means we can possibly lose a lot of logged information that we didn't expect to lose, without even know we lost it.

This commit provides an improvement on that built-in Syslog handler. At startup, it will attempt to determine the minimum MTU between the running process and the Syslog server, or make an educated guess if it can't (or set an upper boundary of 1MB over TCP and Unix sockets). It will then either truncate all messages to come in under that MTU or fragment messages into multiple Syslog messages, based on the configured setting.

This Syslog handler replaces the built-in Syslog handler in the default Server settings.