[MINOR] Add forked process ID for creating deterministic heartbeat files
When using heartbeat files together with `--fork`, you either got errors on shutdown because all processes used the same heartbeat file, or you used `{{pid}}` in the heartbeat file template to create unique heartbeat files, but the file names were non-deterministic and, thus, useless for health checks. This commit enhances forking to send a unique, deterministic forked process ID to each fork child, and a `{{fid}}` replacement in the heartbeat file template to support using that in the heartbeat file name, so that you can have unique heartbeat files per-process that are also useful for healthchecks.