[PATCH] Allow multiple uses of the same stub_action instance
Currently, using the same `stub_action()` instance more than once at the
same times, breaks how the `Client` works, because:
1. It generates infinite recursion when wrapping the `send_request` method.
2. It doesn't completely revert the stub, so the `Client` keeps stubbed.

These two scenarios are fixed by checking if the `stub_action` instance
is already enabled. If it is, the `__enter__` method becomes a no-op.
When it's not enabled, `__exit__` becomos a no-op`.

Added a unit test with the simplest scenario to reproduce the issue.