[MINOR] #204: Add is_caller_error attribute to Error objects
- Implement #204: Add `is_caller_error` attribute to `Error` objects to differentiate server/code errors from caller errors.
- Added `added_in_version` metadata checking to `Server` to ensure we don't return unsupported request attributes to older clients.
- Unhandled exceptions that occur during action execution are now handled as action response errors instead of job response errors, and so are now subject to the regular control flow, such as `continue_on_error`.
- The `ContextVar` compatibility layer now supports `reset` so that we can use it for our metric and error handling in our internal middleware.
- Updated some documentation.
- Fix a couple minor bugs discovered while testing these changes.

Note: `Error.is_caller_error` is always a boolean. The default behavior of `ActionError` is that all of its `Error.is_caller_error` values get set to `True`, but this can be changed to `False`, or set to `None` so that `ActionError` does not change its `Error.is_caller_error` values. The default behavior of `JobError` is that all of its `Error.is_caller_error` values get set to `False`, but this can be changed to `True`, or set  to `None` so that `JobError` does not change its `Error.is_caller_error` values.