[PATCH] Skip Django database connection cleanup in PyTest fixtures
Similar to 037cb2f2f5c717b67264c396654966002ef17c8d, we also need to skip Django database connection cleanup when the local client is created as a PyTest fixture. Note that this may look like a breaking change, but it is not. These fixtures do not get called directly; rather, PyTest calls them and automatically adjusts parameters as necessary, so this is a safe change.