[PATCH] Fix #239: stub_action side_effect now supports mix of types
`stub_action` was not properly popping certain responses from the collection of outstanding responses when those responses were retrieved. As a result, if a `stub_action` call specified a `side_effect`, and one of those side effects was an exception, and other non-exception side effects followed the exception, the exception masked all future side effects. This is now fixed.