[PATCH] Fix #251: Replace use of Django's close_old_connections
Django's built-in `close_old_connections` can fail to close old or unstable connections if an error occurs with one database in a multi-database environment. Additionally, `close_old_connections` is just not really suited for use directly in PySOA. This change moves Django database maintenance tasks into a new module and calls those from the `Server` class, and it replaces the use of `close_old_connections` with a new function that performs extensive error checking.